Move CellEffect to shared package (#122)
* Move client and server CellEffect to shared

* Remove empty CellEffect files

* Move GameType to shared.

* Move Owner to Shared

* Move TargetCardType to shared

* Move CardAttackType to Shared

* Move Target to Shared
