--- Client/src/models/game/GameType.java
@@ -1,5 +1,0 @@
-package models.game;
-
-public enum GameType {
-    KILL_HERO
-}
