+++ Client/src/models/message/Message.java
@@ -7,7 +7,7 @@ import models.account.AccountType;
 import models.card.Card;
 import models.card.ExportedDeck;
 import models.comperessedData.CompressedCard;
-import models.game.GameType;
+import shared.models.game.GameType;
 import shared.models.game.map.Cell;
 
 public class Message {
