+++ Client/src/view/GlobalGameMenu.java
@@ -2,7 +2,7 @@ package view;
 
 
 import controller.MultiPlayerMenuController;
-import models.game.GameType;
+import shared.models.game.GameType;
 
 import java.io.FileNotFoundException;
