+++ Server/src/server/clientPortal/models/comperessedData/CompressedSpell.java
@@ -1,7 +1,8 @@
 package server.clientPortal.models.comperessedData;
 
 import server.dataCenter.models.card.spell.AvailabilityType;
-import server.dataCenter.models.card.spell.Target;
+import shared.models.card.spell.Target;
+
 
 public class CompressedSpell {
     private String spellId;
