+++ Server/src/server/clientPortal/models/message/GameUpdateMessage.java
@@ -1,6 +1,6 @@
 package server.clientPortal.models.message;
 
-import server.gameCenter.models.game.CellEffect;
+import shared.models.game.map.CellEffect;
 
 import java.util.List;
 
@@ -10,7 +10,7 @@ class GameUpdateMessage {
     private final int player2CurrentMP;
     private final List<CellEffect> cellEffects;
 
-    GameUpdateMessage(int turnNumber, int player1CurrentMP,  int player2CurrentMP, List<CellEffect> cellEffects) {
+    GameUpdateMessage(int turnNumber, int player1CurrentMP, int player2CurrentMP, List<CellEffect> cellEffects) {
         this.turnNumber = turnNumber;
         this.player1CurrentMP = player1CurrentMP;
         this.player2CurrentMP = player2CurrentMP;
