+++ Server/src/server/clientPortal/models/message/Message.java
@@ -9,7 +9,9 @@ import server.dataCenter.models.card.Card;
 import server.dataCenter.models.card.ExportedDeck;
 import server.dataCenter.models.card.spell.AvailabilityType;
 import server.gameCenter.models.game.*;
+import shared.models.game.GameType;
 import shared.models.game.map.Cell;
+import shared.models.game.map.CellEffect;
 
 import java.util.List;
 import java.util.Set;
