+++ Server/src/server/clientPortal/models/message/NewGameFields.java
@@ -1,6 +1,6 @@
 package server.clientPortal.models.message;
 
-import server.gameCenter.models.game.GameType;
+import shared.models.game.GameType;
 
 public class NewGameFields {
     private GameType gameType;
