--- Server/src/server/gameCenter/models/game/CellEffect.java
@@ -1,13 +1,0 @@
-package server.gameCenter.models.game;
-
-import shared.models.game.map.Cell;
-
-public class CellEffect {
-    private Cell cell;
-    private boolean positive;
-
-    public CellEffect(Cell cell, boolean positive) {
-        this.cell = cell;
-        this.positive = positive;
-    }
-}
