+++ Shared/src/shared/models/card/spell/CardAttackType.java
@@ -1,20 +1,20 @@
-package server.dataCenter.models.card.spell;
+package shared.models.card.spell;
 
 public class CardAttackType {
     private boolean melee;
     private boolean ranged;
     private boolean hybrid;
 
-    public CardAttackType(CardAttackType attackType) {
-        this.melee = attackType.melee;
-        this.ranged = attackType.ranged;
-        this.hybrid = attackType.hybrid;
-    }
-
     public CardAttackType(boolean melee, boolean ranged, boolean hybrid) {
         this.melee = melee;
         this.ranged = ranged;
         this.hybrid = hybrid;
+    }
+
+    public CardAttackType(CardAttackType attackType) {
+        this.melee = attackType.melee;
+        this.ranged = attackType.ranged;
+        this.hybrid = attackType.hybrid;
     }
 
     public boolean isMelee() {
