+++ Shared/src/shared/models/game/GameType.java
@@ -1,4 +1,5 @@
-package server.gameCenter.models.game;
+package shared.models.game;
+
 
 public enum GameType {
     KILL_HERO
