+++ Shared/src/shared/models/game/map/CellEffect.java
@@ -1,15 +1,19 @@
-package models.game;
-
-import shared.models.game.map.Cell;
+package shared.models.game.map;
 
 public class CellEffect {
     private Cell cell;
     private boolean positive;
 
-    public CellEffect(int row, int column, boolean positive) {
+    public CellEffect(Cell cell, boolean positive) {
+        this.cell = cell;
         this.positive = positive;
-        cell = new Cell(row, column);
     }
+
+    public CellEffect(int row, int column, boolean positive) {
+        cell = new Cell(row, column);
+        this.positive = positive;
+    }
+
 
     public Cell getCell() {
         return cell;
